+++ src/sqrt/utils.ts
@@ -1,19 +1,15 @@
 import { isLeftGreaterInner } from "../compare/utils.ts";
 import { PipeInner } from "../pipe/utils.ts";
-import type { BI } from "../shared/types.ts";
 import type { SqrtInner } from "./types.ts";
 
-const halfInner: BI = [5n, 1];
-const precDef: BI = [1n, 13];
-
-export const sqrtInner: SqrtInner = (input, prec = precDef, max = 100) => {
-  let guess = input;
+export const sqrtInner: SqrtInner = (value, prec = [1n, 13], max = 100) => {
+  let guess = value;
 
   for (let i = 0; i < max; i++) {
     const nextGuess = new PipeInner()
-      .div([input, guess])
+      .div([value, guess])
       .add([guess])
-      .mul([halfInner])
+      .mul([[5n, 1]])
       .calc();
 
     let [bi, fpe] = new PipeInner().sub([nextGuess, guess]).calc();